almost finished with chat flow

+++ backend/services/chat/internal/db/dbservice/queries.go
@@ -75,10 +75,13 @@ const (
 	// ====================================
 	// PRIVATE_CONVERSATIONS
 	// ====================================
-
+	// alex: added casting to least and greatest cause there was an error, as suggested by the clanker, haven't tested it though
 	getOrCreatePrivateConv = `
     INSERT INTO private_conversations (user_a, user_b)
-    VALUES (LEAST($1, $2), GREATEST($1, $2))
+    VALUES (
+		LEAST($1::bigint, $2::bigint),
+		GREATEST($1::bigint, $2::bigint)
+	)
     ON CONFLICT (user_a, user_b) DO UPDATE
     SET updated_at = private_conversations.updated_at
     RETURNING *;
